Restructuration
Made a new base-structure for the project, based on the first one.
The old one is still accessible in src/old/first structuration.
Doxygen added for files, defines, and structs.
Added a basic gestion of events to quit the program.
Added a TODO file at the root to write ongoing works