updated XML parser with working tag reading
It's now possible to create a tag structure from a file.
Tag's name and attributes are readed from the file and the
tag's type is deducted from the slash '/' position in the tag.

Last thing to do is to create a tree from parsed tags. Node's
value reading will be added later, since the TMX map format doesn't
use them.